requirements.txt
aniso8601==8.0.0
click==7.1.2
Flask==1.1.2
Flask-Cors==3.0.9
Flask-RESTful==0.3.8
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
PyMySQL==0.10.1
pytz==2020.1
six==1.15.0
Flask-SQLAlchemy==2.4.4
flask-marshmallow==0.14.0
marshmallow-sqlalchemy==0.23.1
uWSGI==2.0.19.1
Werkzeug==1.0.1
python-dotenv==0.14.0
pyjwt==1.7.1
flask_limiter==1.4
